+++ src/configure/resources/constants.ts
@@ -143,3 +143,6 @@ export const SelectWebApp = 'selectWebApp';
 export const GitHubPat = 'gitHubPat';
 export const SelectFromMultipleWorkSpace = 'selectFromMultipleWorkSpace';
 export const SelectRemoteForRepo = 'selectRemoteForRepo';
+export const VstsRmScmType = 'VSTSRM';
+export const DeploymentResourceAlreadyConfigured = 'DeploymentResourceAlreadyConfigured';
+export const BrowseNotAvailableConfigurePipeline =  'BrowseNotAvailableConfigurePipeline';