+++ src/configure/resources/telemetryKeys.ts
@@ -10,6 +10,10 @@ export class TelemetryKeys {
     public static BrowsePipelineClicked: string = 'browsePipelineClicked';
     public static MultipleWorkspaceFolders: string = 'multipleWorkspaceFolders';
     public static GitFolderExists: string = 'gitFolderExists';
+    public static ScmType: string = 'scmType';
+    public static BrowsedDeploymentCenter = 'openedDeploymentCenter';
+    public static BrowsedExistingPipeline = 'browsedExistingPipeline';
+    public static ClickedConfigurePipeline = 'clickedConfigurePipeline';
 
     // Durations
     public static ExtensionActivationDuration = 'extensionActivationDuration';